Add working persistent drawer
