+++ src/components/Editor.js
@@ -4,18 +4,9 @@ import React, { Component } from 'react';
 import AceEditor from 'react-ace';
 import type { EditorProps } from 'react-ace';
 
-import Drawer from 'material-ui/Drawer';
-import Divider from 'material-ui/Divider';
-import Toolbar from 'material-ui/Toolbar';
-import IconButton from 'material-ui/IconButton';
-import MenuIcon from 'material-ui-icons/Menu';
-import ChevronRightIcon from 'material-ui-icons/ChevronRight';
-import Typography from 'material-ui/Typography';
-import { List, ListItem, ListItemIcon, ListItemText } from 'material-ui/List';
-import InboxIcon from 'material-ui-icons/MoveToInbox';
-
 import 'brace/theme/dracula';
 
+import PersistentDrawer from './PersistentDrawer'
 import styles from './Editor.module.css';
 import errorMarker from '../utils/highlighter';
 import { aceStyleProps } from '../constants';
@@ -30,96 +21,11 @@ type Props = {
 
 type State = {
   code: string,
-  anchor: string,
-  open: boolean,
 };
-
-const drawerWidth = 240;
-
-const styles2 = theme => ({
-  root: {
-    flexGrow: 1,
-  },
-  appFrame: {
-    height: 430,
-    zIndex: 1,
-    overflow: 'hidden',
-    position: 'relative',
-    display: 'flex',
-    width: '100%',
-  },
-  appBar: {
-    position: 'absolute',
-    transition: theme.transitions.create(['margin', 'width'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-  },
-  appBarShift: {
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['margin', 'width'], {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  'appBarShift-left': {
-    marginLeft: drawerWidth,
-  },
-  'appBarShift-right': {
-    marginRight: drawerWidth,
-  },
-  menuButton: {
-    marginLeft: 12,
-    marginRight: 20,
-  },
-  hide: {
-    display: 'none',
-  },
-  drawerPaper: {
-    position: 'relative',
-    width: drawerWidth,
-  },
-  drawerHeader: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'flex-end',
-    padding: '0 8px',
-    ...theme.mixins.toolbar,
-  },
-  content: {
-    flexGrow: 1,
-    backgroundColor: theme.palette.background.default,
-    padding: theme.spacing.unit * 3,
-    transition: theme.transitions.create('margin', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-  },
-  'content-left': {
-    marginLeft: -drawerWidth,
-  },
-  'content-right': {
-    marginRight: -drawerWidth,
-  },
-  contentShift: {
-    transition: theme.transitions.create('margin', {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  'contentShift-left': {
-    marginLeft: 0,
-  },
-  'contentShift-right': {
-    marginRight: 0,
-  },
-});
 
 export default class Editor extends Component<Props, State> {
   state = {
     code: window.initialCode || '',
-    open: false,
-    anchor: 'left',
   };
 
   onEditorLoad = (editor: EditorProps) => {
@@ -139,87 +45,25 @@ export default class Editor extends Component<Props, State> {
     this.setState({ code: newCode });
   };
 
-  handleDrawerOpen = () => {
-    this.setState({ open: true });
-  };
-
-  handleDrawerClose = () => {
-    this.setState({ open: false });
-  };
-
-  handleChangeAnchor = event => {
-    this.setState({
-      anchor: event.target.value,
-    });
-  };
-
   handleClick = () => {
     if (this.state.code.trim()) this.props.runRequest(this.state.code);
   };
 
   render() {
-    const { anchor, open } = this.state;
-
-    const drawer = (
-      <Drawer
-        variant="persistent"
-        anchor={anchor}
-        open={open}
-        classes={{
-          paper: styles2.drawerPaper,
-        }}
-      >
-        <div className={styles2.drawerHeader}>
-          <IconButton onClick={this.handleDrawerClose}>
-            <ChevronRightIcon />
-          </IconButton>
-        </div>
-        <Divider />
-        <List>
-          <div>
-            <ListItem button>
-              <ListItemIcon>
-                <InboxIcon />
-              </ListItemIcon>
-              <ListItemText primary="Inbox" />
-            </ListItem>
-          </div>
-        </List>
-        <Divider />
-      </Drawer>
-    );
-
-    let before = null;
-
-    if (anchor === 'left') {
-      before = drawer;
-    }
+    var aceEditor = (<AceEditor
+      name="code"
+      mode="text"
+      theme="dracula"
+      value={this.state.code}
+      onChange={this.handleChange}
+      onLoad={this.onEditorLoad}
+      {...aceStyleProps}
+      markers={this.getMarkers()}
+    />);
 
     return (
       <div className={styles.editor}>
-        <Toolbar disableGutters={!open}>
-          <IconButton
-            color="inherit"
-            aria-label="open drawer"
-            onClick={this.handleDrawerOpen}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Typography variant="title" color="inherit" noWrap>
-            COCONUT INTERPRETER
-          </Typography>
-        </Toolbar>
-        {before}
-        <AceEditor
-          name="code"
-          mode="text"
-          theme="dracula"
-          value={this.state.code}
-          onChange={this.handleChange}
-          onLoad={this.onEditorLoad}
-          {...aceStyleProps}
-          markers={this.getMarkers()}
-        />
+        <PersistentDrawer aceEditor={aceEditor}/>
       </div>
     );
   }
